requirements.txt
concurrent_log_handler==0.9.28
fastapi==0.116.1
fitz==0.0.1.dev2
lxml==6.0.0
openpyxl==3.1.5
Pillow==11.3.0
pydantic_settings==2.10.1
python_docx==1.2.0
Requests==2.32.5
validators==0.35.0
xlrd==2.0.2
